# Export denominator! and numerator!