Also allow passing pointers into `numerator!`, and change both to not
use `ccall`. This makes it a little bit faster.

Before:

    julia> using Nemo; z = ZZRingElem(); a = QQFieldElem(2, 3);

    julia> @btime Nemo.numerator!($z, $a);
      5.208 ns (0 allocations: 0 bytes)

    julia> @btime Nemo.denominator!($z, $a);
      5.166 ns (0 allocations: 0 bytes)

    julia> a = a^100;

    julia> @btime Nemo.numerator!($z, $a);
      7.083 ns (0 allocations: 0 bytes)

    julia> @btime Nemo.denominator!($z, $a);
      6.750 ns (0 allocations: 0 bytes)

After:

    julia> using Nemo; z = ZZRingElem(); a = QQFieldElem(2, 3);

    julia> @btime numerator!($z, $a);
      4.000 ns (0 allocations: 0 bytes)

    julia> @btime denominator!($z, $a);
      4.000 ns (0 allocations: 0 bytes)

    julia> a = a^100;

    julia> @btime numerator!($z, $a);
      5.875 ns (0 allocations: 0 bytes)

    julia> @btime denominator!($z, $a);
      5.625 ns (0 allocations: 0 bytes)

Resolves #1863

This is a breaking change because older Hecke versions define their own `numerator!` and would be broken by this change (though newer Hecke should be fine -- we'll see).